disco/info: new package containing disco features
Previously we had a disco.Feature type. However, packages may want to
provide their feature as a convenience to be easily registered in the
future. If they do this, we will create import loops between any
packages that are already imported by disco (ie. jid and paging). Moving
features out to its own info package will allow other packages to import
it without bringing in all of the disco package.

Signed-off-by: Sam Whited <sam@samwhited.com>